Remove classes from aliases list
Eloquent and Seeder were used to extend from which is a bad practice.
The SoftDeletingTrait should be imported just like the UserTrait and the RemindableTrait.
Str was also removed because it's just a shortcut for the namespace. People can always re-add it if they like.
I wasn't entirely sure what FormRequest was doing here but I have a feeling it's going to be used for the same reasons as one of the above classes. So I removed it as well.